changed erroneous ints to doubles in cell where member varibles went from ints to doubles

+++ source/Cell.cpp
@@ -1,6 +1,6 @@
 #include "Cell.hpp"
 
-Cell::Cell(bool state,int predDensity, int preyDensity):m_state(state)
+Cell::Cell(bool state,double predDensity, double preyDensity):m_state(state)
 {
 	//if the cell is wet it can't contain any predators or prey
 	if(m_state == false)
@@ -16,11 +16,11 @@ Cell::Cell(bool state,int predDensity, int preyDensity):m_state(state)
 }
 Cell::~Cell(){}
 
-int Cell::getPredDensity() const {return m_predDensity;}
-int Cell::getPreyDensity() const {return m_preyDensity;}
-bool Cell::getState() const	     {return m_state;}
+double Cell::getPredDensity() const {return m_predDensity;}
+double Cell::getPreyDensity() const {return m_preyDensity;}
+bool   Cell::getState()       const	{return m_state;}
 
-void Cell::setPredDensity(int predDensity){m_predDensity = predDensity;}
-void Cell::setPreyDensity(int preyDensity){m_preyDensity = preyDensity;}
-void Cell::setState(bool state) 		  {m_state = state;}
+void Cell::setPredDensity(double predDensity){m_predDensity = predDensity;}
+void Cell::setPreyDensity(double preyDensity){m_preyDensity = preyDensity;}
+void Cell::setState(bool state) 		     {m_state = state;}
